gcp_weather_pulse/fetcher/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
PyYAML==6.0.2
google-cloud-storage==2.18.2
google-cloud-pubsub==2.21.5